# Reduced the size of the RateMap