--- a/debian/changelog
+++ b/debian/changelog
@@ -1,6 +1,6 @@
   [Michele Simionato]
-  * Reduced the size of the global RateMap in case of cluster sources
-    (USA model)
+  * Optimized "building cmakers" and the RateMap in the case of clusters
+    of sources (260x for the latest USA model)
 
   [Christopher Brooks]
   * Added BSSA14 site term capability to Modifiable GMPE + unit tests