… openquake/calculators/tests/logictree_test.py
… openquake/qa_tests_data/logictree/case_05/expected/curve-mean-bis.csv
… openquake/qa_tests_data/logictree/case_05/expected/curve-mean.csv
… utils/profile_ctxs.py