Closes https://github.com/gem/oq-engine/issues/10823. The global RateMap for USA is now 13.7 GB instead of 3.5 TB i.e. 260x smaller and `get_cmakers` is 260x faster which makes a huge difference for single site calculations (from 18 mins to 4 secs for the get_cmakers part). Also, the postclassical becomes 3x faster and "total save_rates" 18x faster.
Here are the numbers for 10% of the USA model with 1600 realizations:
```
# before
| calc_1760, maxmem=380.7 GB | time_sec  | memory_mb | counts     |
|----------------------------+-----------+-----------+------------|
| total classical            | 525_444   | 854.4     | 1_196      |
| get_poes                   | 249_414   | 0.0       | 17_822_389 |
| computing mean_std         | 103_997   | 0.0       | 530_335    |
| total postclassical        | 19_711    | 418.6     | 1_771      |
| combine pmaps              | 15_802    | 0.0       | 23_429     |
| reading rates              | 3_890     | 418.5     | 1_771      |
| total save_rates           | 14_868    | 0.2266    | 81         |
| ClassicalCalculator.run    | 7_979     | 2_750     | 1          |

| /home/michele/oqdata/calc_1760.hdf5 | 308.39 MB |
| /home/michele/tmp/michele/calc_1760 | 21.11 GB  |

# after
| calc_1765, maxmem=409.9 GB | time_sec  | memory_mb | counts     |
|----------------------------+-----------+-----------+------------|
| total classical            | 487_763   | 880.9     | 1_195      |
| get_poes                   | 247_661   | 0.0       | 17_818_238 |
| computing mean_std         | 103_732   | 0.0       | 529_951    |
| total postclassical        | 6_009     | 94.8750   | 27         |
| reading rates              | 5_847     | 94.8750   | 27         |
| total save_rates           | 816.2     | 0.1875    | 77         |
| ClassicalCalculator.run    | 4_166     | 1_384     | 1          |

| /home/michele/oqdata/calc_1765.hdf5 | 304.41 MB |
| /home/michele/tmp/michele/calc_1765 | 23.1 GB   |
```